pyproject.toml
[project]
name = "comfyui_yvann-nodes"
description = "Audio Reactive nodes for AI animations 🔊 Analyze audio, extract drums, bass, vocals. Compatible with IPAdapter, ControlNets, AnimateDiff... Generate reactive masks and weights. Create audio-driven visuals. Produce weight graphs and audio masks. Ideal for music videos and reactive animations. Features audio scheduling and waveform analysis"
version = "2.0.5"
license = {file = "LICENCE"}
requires-python = ">=3.9"
dependencies = ["openunmix", "numpy", "torch", "matplotlib", "pillow", "scipy", "torchaudio", "termcolor"]

[project.urls]
Repository = "https://github.com/yvann-ba/ComfyUI_Yvann-Nodes"
#  Used by Comfy Registry https://comfyregistry.org

[tool.comfy]
PublisherId = "yvann"
DisplayName = "ComfyUI_Yvann-Nodes"
Icon = ""
